More rationalizing of Trading API and impl class locations
Converted BalanceInfo to interface and provided default impl in
Exchange Adapter project.
