+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/BitfinexExchangeAdapter.java
@@ -26,6 +26,7 @@ package com.gazbert.bxbot.exchanges;
 import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.BalanceInfoImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.OpenOrderImpl;
@@ -404,7 +405,7 @@ public final class BitfinexExchangeAdapter extends AbstractExchangeAdapter imple
             }
 
             // 2nd arg of BalanceInfo constructor for reserved/on-hold balances is not provided by exchange.
-            return new BalanceInfo(balancesAvailable, new HashMap<>());
+            return new BalanceInfoImpl(balancesAvailable, new HashMap<>());
 
         } catch (ExchangeNetworkException | TradingApiException e) {
             throw e;
