+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/BitstampExchangeAdapter.java
@@ -26,6 +26,7 @@ package com.gazbert.bxbot.exchanges;
 import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.BalanceInfoImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.OpenOrderImpl;
@@ -384,7 +385,7 @@ public final class BitstampExchangeAdapter extends AbstractExchangeAdapter imple
             balancesOnOrder.put("LTC", balances.ltc_reserved);
             balancesOnOrder.put("XRP", balances.xrp_reserved);
 
-            return new BalanceInfo(balancesAvailable, balancesOnOrder);
+            return new BalanceInfoImpl(balancesAvailable, balancesOnOrder);
 
         } catch (ExchangeNetworkException | TradingApiException e) {
             throw e;
