+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/GdaxExchangeAdapter.java
@@ -27,6 +27,7 @@ import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.BalanceInfoImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.OpenOrderImpl;
@@ -424,7 +425,7 @@ public final class GdaxExchangeAdapter extends AbstractExchangeAdapter implement
                     balancesAvailable.put(gdaxAccount.currency, gdaxAccount.available);
                     balancesOnHold.put(gdaxAccount.currency, gdaxAccount.hold);
                 }
-                return new BalanceInfo(balancesAvailable, balancesOnHold);
+                return new BalanceInfoImpl(balancesAvailable, balancesOnHold);
 
             } else {
                 final String errorMsg = "Failed to get your wallet balance info from exchange. Details: " + response;
