+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/GeminiExchangeAdapter.java
@@ -27,6 +27,7 @@ import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.BalanceInfoImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.OpenOrderImpl;
@@ -454,7 +455,7 @@ public final class GeminiExchangeAdapter extends AbstractExchangeAdapter impleme
                     .forEach(accountBalance -> balancesAvailable.put(accountBalance.currency, accountBalance.available));
 
             // 2nd arg of BalanceInfo constructor for reserved/on-hold balances is not provided by exchange.
-            return new BalanceInfo(balancesAvailable, new HashMap<>());
+            return new BalanceInfoImpl(balancesAvailable, new HashMap<>());
 
         } catch (ExchangeNetworkException | TradingApiException e) {
             throw e;
