+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/KrakenExchangeAdapter.java
@@ -27,6 +27,7 @@ import com.gazbert.bxbot.exchange.api.AuthenticationConfig;
 import com.gazbert.bxbot.exchange.api.ExchangeAdapter;
 import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
+import com.gazbert.bxbot.exchanges.trading.api.impl.BalanceInfoImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.OpenOrderImpl;
@@ -685,7 +686,7 @@ public final class KrakenExchangeAdapter extends AbstractExchangeAdapter impleme
                         }
 
                         // 2nd arg of BalanceInfo constructor for reserved/on-hold balances is not provided by exchange.
-                        return new BalanceInfo(balancesAvailable, new HashMap<>());
+                        return new BalanceInfoImpl(balancesAvailable, new HashMap<>());
 
                     } else {
